Create Tree Cordinates
Problem statement:-
We consider metric space to be a pair, , where  is a set and  such that the following conditions hold:
where  is the distance between points  and . Let's define the product of two metric spaces, , to be  such that: where , .
So, it follows logically that  is also a metric space. We then define squared metric space, , to be the product of a metric space multiplied with itself: .

For example, , where  is a metric space. , where .

In this challenge, we need a tree-space. You're given a tree, , where  is the set of vertices and  is the set of edges. Let the function  be the distance between two vertices in tree  (i.e.,  is the number of edges on the path between vertices  and ). Note that  is a metric space.
You are given a tree, , with  vertices, as well as  points in . Find and print the distance between the two furthest points in this metric space!